+++ mempool/src/shared_mempool/peer_manager.rs
@@ -1,12 +1,18 @@
 // Copyright (c) The Libra Core Contributors
 // SPDX-License-Identifier: Apache-2.0
 
-use crate::counters;
+use crate::{
+    counters,
+    logging::{LogEntry, LogEvent, LogSchema},
+    network::MempoolSyncMsg,
+    shared_mempool::types::{notify_subscribers, SharedMempool, SharedMempoolNotification},
+};
 use itertools::Itertools;
 use libra_config::{
     config::{PeerNetworkId, UpstreamConfig},
     network_id::NetworkId,
 };
+use libra_logger::prelude::*;
 use netcore::transport::ConnectionOrigin;
 use rand::seq::SliceRandom;
 use serde::{Deserialize, Serialize};
@@ -16,6 +22,7 @@ use std::{
     sync::Mutex,
     time::Instant,
 };
+use vm_validator::vm_validator::TransactionValidation;
 
 /// stores only peers that receive txns from this node
 pub(crate) type PeerInfo = HashMap<PeerNetworkId, PeerSyncState>;
@@ -128,6 +135,154 @@ impl PeerManager {
             state.is_alive = false;
         }
         self.update_failover();
+    }
+
+    pub fn is_backoff_mode(&self, peer: &PeerNetworkId) -> bool {
+        self.peer_info
+            .lock()
+            .expect("failed to acquire peer info lock")
+            .get(peer)
+            .expect("missing peer info for peer")
+            .broadcast_info
+            .backoff_mode
+    }
+
+    pub fn execute_broadcast<V>(
+        &self,
+        peer: PeerNetworkId,
+        scheduled_backoff: bool,
+        smp: &mut SharedMempool<V>,
+    ) where
+        V: TransactionValidation,
+    {
+        // start timer for tracking broadcast latency
+        let start_time = Instant::now();
+        let peer_manager = &smp.peer_manager;
+
+        let (timeline_id, retry_txns_id, next_backoff) = if peer_manager.is_picked_peer(&peer) {
+            let state = peer_manager.get_peer_state(&peer);
+            let next_backoff = state.broadcast_info.backoff_mode;
+            if state.is_alive {
+                (
+                    state.timeline_id,
+                    state
+                        .broadcast_info
+                        .total_retry_txns
+                        .into_iter()
+                        .collect::<Vec<_>>(),
+                    next_backoff,
+                )
+            } else {
+                return;
+            }
+        } else {
+            return;
+        };
+
+        // It is possible that a broadcast was scheduled as non-backoff before an ACK received after the
+        // broadcast scheduling turns on backoff mode
+        // If this is the case, ignore this schedule and wait till next broadcast scheduled as backoff
+        if !scheduled_backoff && next_backoff {
+            return;
+        }
+
+        // craft batch of txns to broadcast
+        let mut mempool = smp
+            .mempool
+            .lock()
+            .expect("[shared mempool] failed to acquire mempool lock");
+
+        // first populate batch with retriable txns, to prioritize resending them
+        let retry_txns = mempool.filter_read_timeline(retry_txns_id);
+        // pad the batch with new txns from fresh timeline read, if batch has space
+        let (new_txns, new_timeline_id) = if retry_txns.len() < smp.config.shared_mempool_batch_size
+        {
+            mempool.read_timeline(
+                timeline_id,
+                smp.config.shared_mempool_batch_size - retry_txns.len(),
+            )
+        } else {
+            (vec![], timeline_id)
+        };
+
+        if new_txns.is_empty() && retry_txns.is_empty() {
+            return;
+        }
+
+        // read first tx in timeline
+        let earliest_timeline_id = mempool
+            .read_timeline(0, 1)
+            .0
+            .get(0)
+            .expect("empty timeline")
+            .0;
+        // don't hold mempool lock during network send
+        drop(mempool);
+
+        // combine retry_txns and new_txns into batch
+        let mut all_txns = retry_txns
+            .into_iter()
+            .chain(new_txns.into_iter())
+            .collect::<Vec<_>>();
+        all_txns.truncate(smp.config.shared_mempool_batch_size);
+        let batch_timeline_ids = all_txns.iter().map(|(id, _txn)| *id).collect::<Vec<_>>();
+        let batch_txns = all_txns
+            .into_iter()
+            .map(|(_id, txn)| txn)
+            .collect::<Vec<_>>();
+
+        let network_sender = smp
+            .network_senders
+            .get_mut(&peer.network_id())
+            .expect("[shared mempool] missing network sender");
+
+        let batch_id = BatchId(timeline_id, new_timeline_id);
+        let request_id = if let Ok(bytes) = lcs::to_bytes(&batch_id) {
+            bytes
+        } else {
+            // TODO log this
+            return;
+        };
+
+        let txns_ct = batch_txns.len();
+        if let Err(e) = network_sender.send_to(
+            peer.peer_id(),
+            MempoolSyncMsg::BroadcastTransactionsRequest {
+                request_id,
+                transactions: batch_txns,
+            },
+        ) {
+            counters::NETWORK_SEND_FAIL
+                .with_label_values(&[counters::BROADCAST_TXNS])
+                .inc();
+            error!(
+                LogSchema::event_log(LogEntry::BroadcastTransaction, LogEvent::NetworkSendFail)
+                    .peer(&peer)
+                    .error(&e.into())
+            );
+        } else {
+            let broadcast_time = Instant::now();
+            let peer_id = &peer.peer_id().to_string();
+            counters::SHARED_MEMPOOL_TRANSACTION_BROADCAST
+                .with_label_values(&[peer_id])
+                .observe(txns_ct as f64);
+            counters::SHARED_MEMPOOL_PENDING_BROADCASTS_COUNT
+                .with_label_values(&[peer_id])
+                .inc();
+            peer_manager.update_peer_broadcast(
+                peer,
+                batch_id,
+                batch_timeline_ids,
+                new_timeline_id,
+                earliest_timeline_id,
+                broadcast_time,
+            );
+            notify_subscribers(SharedMempoolNotification::Broadcast, &smp.subscribers);
+            let broadcast_latency = start_time.elapsed();
+            counters::SHARED_MEMPOOL_BROADCAST_LATENCY
+                .with_label_values(&[peer_id])
+                .observe(broadcast_latency.as_secs_f64());
+        }
     }
 
     // updates the peer chosen to failover to if all peers in the primary upstream network are down
